add skeleton query handler to get 'commitment agreement's + rename fodler/namespace for existing query from GetAgreement -> GetProviderAgreement (to match the query handler/request/response names) + rename model for consistency + general wip
