+++ src/SFA.DAS.ProviderApprenticeshipsService.Application.UnitTests/SFA.DAS.ProviderApprenticeshipsService.Application.UnitTests.csproj
@@ -150,7 +150,7 @@
     <Compile Include="Commands\UpsertRegisteredUser\WhenValidatingCommand.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Queries\ApprenticeshipSearch\WhenSubmittingApprenticeshipSearch.cs" />
-    <Compile Include="Queries\GetAgreement\WhenIGetAgreement.cs" />
+    <Compile Include="Queries\GetProviderAgreement\WhenIGetProviderAgreement.cs" />
     <Compile Include="Queries\GetApprenticeshipDataLockSummary\WhenGettingApprenticeshipDataLockSummary.cs" />
     <Compile Include="Queries\GetApprenticeshipPriceHistory\WhenGettingApprenticeshipPriceHistory.cs" />
     <Compile Include="Queries\GetBulkUploadFile\WhenIGetBulkUploadFile.cs" />
