+++ src/SFA.DAS.ProviderApprenticeshipsService.Application/Queries/GetCommitmentAgreements/GetCommitmentAgreementsQueryRequest.cs
@@ -1,0 +1,8 @@
+﻿using MediatR;
+
+namespace SFA.DAS.ProviderApprenticeshipsService.Application.Queries.GetCommitmentAgreements
+{
+    public class GetCommitmentAgreementsQueryRequest : IAsyncRequest<GetCommitmentAgreementsQueryResponse>
+    {
+    }
+}
