+++ src/SFA.DAS.ProviderApprenticeshipsService.Application/Queries/GetCommitmentAgreements/GetCommitmentAgreementsQueryResponse.cs
@@ -1,0 +1,6 @@
+﻿namespace SFA.DAS.ProviderApprenticeshipsService.Application.Queries.GetCommitmentAgreements
+{
+    public class GetCommitmentAgreementsQueryResponse
+    {
+    }
+}
