+++ src/SFA.DAS.ProviderApprenticeshipsService.Application/Queries/GetProviderAgreement/GetProviderAgreementQueryHandler.cs
@@ -2,25 +2,19 @@
 using System.Linq;
 using System.Threading.Tasks;
 using MediatR;
-
 using SFA.DAS.ProviderApprenticeshipsService.Domain;
 using SFA.DAS.ProviderApprenticeshipsService.Domain.Data;
 using SFA.DAS.ProviderApprenticeshipsService.Infrastructure.Configuration;
 
-namespace SFA.DAS.ProviderApprenticeshipsService.Application.Queries.GetAgreement
+namespace SFA.DAS.ProviderApprenticeshipsService.Application.Queries.GetProviderAgreement
 {
     public class GetProviderAgreementQueryHandler : IAsyncRequestHandler<GetProviderAgreementQueryRequest, GetProviderAgreementQueryResponse>
     {
         private readonly IAgreementStatusQueryRepository _agreementStatusQueryRepository;
-
         private readonly ProviderApprenticeshipsServiceConfiguration _configuration;
 
         public GetProviderAgreementQueryHandler(IAgreementStatusQueryRepository agreementStatusQueryRepository, ProviderApprenticeshipsServiceConfiguration configuration)
         {
-            if (agreementStatusQueryRepository == null)
-                throw new ArgumentNullException(nameof(agreementStatusQueryRepository));
-            if(configuration == null)
-                throw new ArgumentNullException(nameof(configuration));
             _agreementStatusQueryRepository = agreementStatusQueryRepository;
             _configuration = configuration;
         }
