+++ src/SFA.DAS.ProviderApprenticeshipsService.Web/Models/Agreement/AgreementsViewModel.cs
@@ -1,12 +1,9 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Web;
+﻿using System.Collections.Generic;
 
 namespace SFA.DAS.ProviderApprenticeshipsService.Web.Models.Agreement
 {
     public class AgreementsViewModel
     {
-        public IEnumerable<CohortAgreement> CohortAgreements;
+        public IEnumerable<CommitmentAgreement> CommitmentAgreements;
     }
 }
