+++ src/SFA.DAS.ProviderApprenticeshipsService.Web/Models/Agreement/CommitmentAgreement.cs
@@ -1,7 +1,7 @@
 ﻿
 namespace SFA.DAS.ProviderApprenticeshipsService.Web.Models.Agreement
 {
-    public class CohortAgreement
+    public class CommitmentAgreement
     {
         public string OrganisationName { get; set; }    // LegalEntityName
         public string CohortID { get; set; }            // HashedCommitmentId
