+++ src/SFA.DAS.ProviderApprenticeshipsService.Web/SFA.DAS.ProviderApprenticeshipsService.Web.csproj
@@ -479,7 +479,7 @@
     </Compile>
     <Compile Include="Models\AcknowledgementViewModel.cs" />
     <Compile Include="Models\Agreement\AgreementsViewModel.cs" />
-    <Compile Include="Models\Agreement\CohortAgreement.cs" />
+    <Compile Include="Models\Agreement\CommitmentAgreement.cs" />
     <Compile Include="Models\ApprenticeshipDetailsViewModel.cs" />
     <Compile Include="Models\AgreementNotSignedViewModel.cs" />
     <Compile Include="Models\ApprenticeshipFiltersViewModel.cs" />
