+++ src/SFA.DAS.ProviderApprenticeshipsService.Web/Views/Account/Index.cshtml
@@ -43,7 +43,7 @@
         <div class="grid-row">
             <div class="column-one-half">
                 <h2 class="heading-medium">
-                    <a href="@Url.Action("Cohorts", "Commitment", new {providerId = Model.ProviderId})">Organisations and agreements</a>
+                    <a href="@Url.Action("", "Agreement", new {providerId = Model.ProviderId})">Organisations and agreements</a>
                 </h2>
                 <p>Find and view organisations and their agreement IDs.</p>
                 <hr />
