+++ src/SFA.DAS.ProviderApprenticeshipsService.Web/Views/Agreement/Agreements.cshtml
@@ -24,7 +24,7 @@
                 </tr>
             </thead>
             <tbody>
-                @foreach (var cohortAgreement in Model.CohortAgreements)
+                @foreach (var cohortAgreement in Model.CommitmentAgreements)
                 {
                     <tr>
                         <td>
